refactor: delete integration/ — move Go API tests to core/api, drop gateway e2e dupes
integration/e2e_test.go → core/api/e2e_test.go: 4 tests exercising the
full Go API stack (Create→Resolve, Update→Resolve, ChainedUpdates with
Lineage, InsertDelete). These are Go-level API tests, not HTTP e2e.

integration/gateway_e2e_test.go → deleted entirely: all 5 tests
(FullLifecycle, UpdateAndResolve, ManagedGraphHeadLifecycle, Health,
BatchUpdate) are already covered by gateway/gateway_test.go with more
comprehensive assertions and proper error code checks.

True deployment-level E2E (latency, proof size, rewrite amplification)
belongs in eval/ as benchmark tests, not in integration/.

Authored-By: Bernard <vitalikzhou@gmail.com>
